feat(kernel): implement SMP scheduler, network stack and core improvements
Add comprehensive kernel updates including:
- Per-CPU run queue scheduler with SMP affinity and CPU hotplug
- Full IPv4/IPv6, TCP/UDP, UNIX socket networking with network namespaces
- POSIX security capabilities, UID/GID handling and syscall filtering
- VFS fixes for uid/gid tracking across all filesystems
- Kernel messaging, panic recovery and lockdep/RCU debugging tools
- Updated block layer, ATA driver and memory management subsystems
- TTY/PTY and pipe implementation improvements
- Added automated test and QEMU run scripts

+++ os/src/include/sys/types.h
@@ -10,8 +10,8 @@ typedef long long uid_t;
 typedef long long gid_t;
 typedef unsigned int mode_t;
 typedef long long time_t;
-typedef long long intptr_t;
-typedef unsigned long long uintptr_t;
+typedef long intptr_t;
+typedef unsigned long uintptr_t;
 #endif
 
 #endif
